+++ internal/otlp_wrapper.go
@@ -18,6 +18,7 @@ import (
 	otlpcollectorlog "go.opentelemetry.io/collector/internal/data/protogen/collector/logs/v1"
 	otlpcollectormetrics "go.opentelemetry.io/collector/internal/data/protogen/collector/metrics/v1"
 	otlpcollectortrace "go.opentelemetry.io/collector/internal/data/protogen/collector/trace/v1"
+	otlptrace "go.opentelemetry.io/collector/internal/data/protogen/trace/v1"
 )
 
 // MetricsWrapper is an intermediary struct that is declared in an internal package
@@ -48,6 +49,30 @@ func TracesToOtlp(mw TracesWrapper) *otlpcollectortrace.ExportTraceServiceReques
 
 func TracesFromOtlp(req *otlpcollectortrace.ExportTraceServiceRequest) TracesWrapper {
 	return TracesWrapper{req: req}
+}
+
+// TracesCompatibilityChanges performs backward compatibility conversion of Span Status code according to
+// OTLP specification as we are a new receiver and sender (we are pushing data to the pipelines):
+// See https://github.com/open-telemetry/opentelemetry-proto/blob/59c488bfb8fb6d0458ad6425758b70259ff4a2bd/opentelemetry/proto/trace/v1/trace.proto#L239
+// See https://github.com/open-telemetry/opentelemetry-proto/blob/59c488bfb8fb6d0458ad6425758b70259ff4a2bd/opentelemetry/proto/trace/v1/trace.proto#L253
+func TracesCompatibilityChanges(req *otlpcollectortrace.ExportTraceServiceRequest) {
+	for _, rss := range req.ResourceSpans {
+		for _, ils := range rss.InstrumentationLibrarySpans {
+			for _, span := range ils.Spans {
+				switch span.Status.Code {
+				case otlptrace.Status_STATUS_CODE_UNSET:
+					if span.Status.DeprecatedCode != otlptrace.Status_DEPRECATED_STATUS_CODE_OK {
+						span.Status.Code = otlptrace.Status_STATUS_CODE_ERROR
+					}
+				case otlptrace.Status_STATUS_CODE_OK:
+					// If status code is set then overwrites deprecated.
+					span.Status.DeprecatedCode = otlptrace.Status_DEPRECATED_STATUS_CODE_OK
+				case otlptrace.Status_STATUS_CODE_ERROR:
+					span.Status.DeprecatedCode = otlptrace.Status_DEPRECATED_STATUS_CODE_UNKNOWN_ERROR
+				}
+			}
+		}
+	}
 }
 
 // LogsWrapper is an intermediary struct that is declared in an internal package